KafkaConsumer/TwitterLLM/requirements.txt
kafka-python==2.0.2
pymongo==4.6.3
requests==2.31.0
python-dotenv==1.0.1
google-generativeai==0.5.0
